Notebook to calculate the vesseldiameter et al.
We take the datasets prepared in the 'DisplayOverviewScans' notebook and
treat them further.
First we do a very simple gray-value threshold based segmentation and
then calculate the medial axis of the vessels.
This is done by filling the hollow bones (to be improved), removing the
filled bones from the segmented data (via a `tophat` filtering) and then
calculating both the skeleton and euclidean distance transformation.
By multiplying the skeleton and edt, we get an approximation of the
medial axis transformation (which doesn't work for the 3D case with
skimage).

Now we need to analyze/plot this for each dataset...